Feat: Add command for I2C/I3C and sensor Mod: Update Makefile and Kconfig for all IP Fix: i3c data reverse

+++ cmd/diag/hal/nvme/ax_nvme.h
@@ -29,7 +29,7 @@
 /*
  * CAP
  */
-#define NVME_CAP_MQES(cap)      ((u16)((cap) & 0xffff))
+#define NVME_CAP_MQES(cap)      ((u32)((cap) & 0xffff))
 #define NVME_CAP_CQR(cap)       (((cap) >> 16) & 0x1)
 #define NVME_CAP_AMS(cap)       (((cap) >> 17) & 0x3)
 #define NVME_CAP_TO(cap)        (((cap) >> 24) & 0xff)
